refactor(iap): drop the hand-rolled verifier, keep Apple's library
Apple's app-store-server-library is verified working on a production
Base44 app: a real sandbox TEST notification was delivered to a backend
function and handled correctly. That was the only open question about
running it on Cloudflare Workers, so the fallback is no longer earning
its keep.

Removes the hand-rolled implementation — the DER reader, X.509 parser,
ECDSA helpers, chain walker and JWS parser — along with the `verifier`
config option and the two-implementation test matrix. About 1,300 lines
of security-critical code we no longer own. Apple's library is also
stricter than what it replaced: it checks the intermediate's CA basic
constraint, which the hand-rolled chain walker did not.

Two behavioural differences are now covered by tests rather than left
implicit. Apple identifies the app by bundleId on a transaction and
checks appAppleId only on a notification — the bundleId identifies the
app on its own, so nothing is lost. And a payload with no signedDate is
dated against the current time rather than refused.

Kept: the pinned Apple roots (the library takes trust anchors as a
constructor argument), the decoded payload types, the environment parse,
and the base64 codec.
